vmd: split base-comparable write benchmarks and add trust-stamp pressure bench
